Initial syscalls.h file
Libc has been a nightmare so we are transitioning to only riscv-unknown-elf compiler support and opting to include our own system call implementations.
Seeing as the system call implementations simply need to load register a7 with the code and the rest is handled the same way any function call is (ie. argument 1 goes to a0, 2 to a1, and so on)...
all of the system call implementations (in `common/syscalls.h`) take in the correct arguments and then manually insert the assembly for the a7 reg & ecall instruction.
The rest of the implementations are handled entirely inside Rev.

It is worth noting that despite many syscalls being included in syscalls.h, not all have corresponding implementations inside of Rev (See `src/RevProc.cc` for ECALL_<syscall_name> functions implementing each or simply returning immediately to prevent sim failures.)