fixing bug where all system calls return 0
Because the function definitions in syscalls.h all originally were returning 0, the `a0` register (function return argument) would be overwritten by 0 everytime.

Now, there is an additional mv instruction which places the value in a0 (after the ecall) into a variable `rc` and that is what is returned.
This *might* half-work if we just didn't return anything but then it wouldn't be possible to write code that depends on the return value of a system call (ie. fork)
because you wouldn't be able to access the pid that's returned from a void function