Remove all useless casts (-Wuseless-cast)
Simplify atomic code:
- Compress implementation of 9 atomic RV32/RV64 instructions into one template
- Iterate through AMO flags in compact tables, instead of long repetitive if-else
  code which is error-prone

Reword comments to not say "default constructor" or "default destructor" unless
it is actually declared as "= default". A user-defined constructor/destructor is
not a "default" one.

Prefer function-style casts Type(expr) to C-style casts (Type)(expr) for brevity,
although static_cast or reintrepret_cast should ideally be used.

Rearrange code to make it more readable (such as in the decoding of instructions).