Revert changes to the Pan* files, except for adding spaces after commas where they were missing.
Change CMakeLists.txt to disable the -Wnon-virtual-dtor warning, since SST gets
that warning, and we either have to change the Pan* files to #include "SST.h"
which suppresses it, or else we have to manually add the suppression #pragmas
to the Pan* files. I rarely see someone delete a derived class from a base
class pointer, so this is not a big issue.