Generalize RevRand to support mixed-type (min, max), and to automatically use floating-point or integer distributions for the common (min, max) type.
Initialize the thread-local RNG inside of a function outside of the RNG
template, so that each thread has only one RNG engine, but it can be
applied to multiple distributions with different return types (different
template instantiations). By putting the initialization inside a function
rather than an inline global variable, the initialization occurs on-demand
and there is no initialization overhead every time a thread is created.