Replace the ResetInst() function with default initializers in RevInst. RevInst will always be initialized now. This gets rid of valgrind uninitialized warnings, but valgrind did not point out the right culprit, which was DecodeCompressed(), which was returning decodings with uninitialized data.
Replace "return TInst;" sentinel values with "return RevInst{};" which
is default-initialized as resetted. The TInst variable was not used
except for the final return statement, which should never be reached.

Make RevInstDefaults use "static constexpr" initializers, so that they
can be optimized as compile-time instead of run-time constants.

Simply static_assert for RevRegFile and RevInst.