REV: Add clock_gettime syscall
Wire clock_gettime to the cycle counter to be able
to use it to measure time intervals in various
segments of workloads. Since REV core is not aware
of the clock it is running at we have to propagate
timeConverter to each of the cores in RevProc, to
convert cycles to real time. Clockid argument is not
supported - we simply ignore it.

Add a test with usage example of the "new" syscall.