Fix the decoding of Reg-4 instructions
Fix the decoding of Reg-4 instructions (`fmadd`, `fmsub`, `fnmadd`, `fnmsub`)
by renaming `Funct7` to `Funct2or7` and filling it in with `Funct2` during
Reg-4 instructions, because in Reg-4 instruction encodings, `Funct7` overloads
`rs3:Funct2`, where `rs3` is the third source register and `Funct2` is the
precision. `Funct2or7` represents either `Funct2` in Reg-4 instructions, or
`Funct7` in all other non-compressed instructions. It is distinct from the
`Funct2` field used in compressed instructions, whose handling is unchanged by
this.

Previously, the `Funct2` field in Reg-4 instructions was ignored, causing
`fmadd.d` and similar double-precision instructions to fail.

There was a 'fmt` field for floating-point format (precision) in `RevInst`, but
it was never filled or used, and using it instead of `Funct2or7` would prevent
the precision from being used during instruction lookup. So the `fmt` field has
been removed.

Also templatize (`fmadd`, `fmsub`, `fnmadd`, `fnmsub`) so that the same code is
used for single and double precisions.

Add a test case which fails before this change, and passes now.