Fix FCVT instructions
- The instruction table had a wrong entry for fcvt.d.s, and the fcvtsd and
fcvtds Execution functions were implemented backwards.

- More FCVT tests were added.

- FMV and FST instructions need to load a FP register in raw mode instead of
unboxing the NaN and verifying that the upper bits are all 1 (see section 12.2
of the unprivleged spec). A new raw mode FMV_FS template parameter has been
added to RevRegFile::GetFP<...>, but it defaults to false (the old behavior).
It's only used when loading a FP register for a FMV, FSW or FSD.

- Some compressed decoding functions still had comments which said "Removed -
set in decode".