Use reference count of the number of memory requests to a RevMemOp instead of a "split request" count and a linear search of outstanding requests to a RevMemOp.
The request reference count is incremented when am outstanding SST memory
request refers to a RevMemOp, and is decremented in handleResp() when a SST
memory request response is handled. When the reference count == 0 in
handleResp(), it indicates all parts of a split memory request have completed.

This reference count is independent of the shared_ptr reference count pointing
to RevMemOp, which is used for memory management instead of detecting
completion of all split requests of a RevMemOp, and may have other references
besides the SST memory requests.