Remove the AMOTable as unnecessary and slow.
We already track outstanding memory requests which refer to RevMemOp
operations, which already has flags which tell whether the operation is atomic,
so we do not need to track atomic operations in a separate AMOTable, whose
searching and management adds linear search time to handleResp() for handling
memory responses. We only have to test the RevMemOp flags to check if the
memory operation is atomic, and handle it accordingly in the memory response.

Add move functions to move MemReq from reader to writer in atomic operations,
so that the MarkLoadComplete() std::function is moved instead of copied.