Remove MemOp::MemOpAMO, since it is not needed, and RevFlagAtomic( flags ) tells whether an operation is atomic.
Fix the handling of sign-/zero-extension and the handling of MarkLoadComplete()
in atomic operations.

Mark the Load of an atomic operation complete as soon as the read completes, the
atomic operation sets the destination register value, and the write is sent.

There is no need to wait for the write to complete before marking the load as
complete, because the register's result value has been set. Acquire and Release
flags control how later memory operations are ordered w.r.t. the write operation.